Remove `psr/container:^2` support, since our implementation does not respect its interfaces
Reverts #72
Fixes #74

Because of our default container implementation not having return type declarations,
and because our container not being `final` (and therefore not being able to change
this without BC breaks), we cannot support `psr/container:^2` for now.

In #72 we erroneously introduced `psr/container:^2` support, but our CI did never
verify this, because our constraints do not allow for its installation, and therefore
we never verified its compatibility.

This means that `3.9.0` will effectively be preferred over `3.9.1`, if people
depend on `psr/container:^2`, but sadly, such is the problem with already tagged
releases and dependency SAT mechanisms.